requirements.txt
numpy
scipy>=1.8.0
pytest>=7.1.1
nexusformat>=0.7.4
fast-histogram>=0.10
pyevtk>=1.5.0
setuptools<58.0.0
transformations
h5py
pyyaml
pandas
tifffile
fabio
pyFAI
schema
concurrent-log-handler